move endpoint get_details v2 into separate file
